requirements.txt
streamlit==1.30.0
pandas==2.1.3
plotly==5.18.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0 